+++ streaming/base/util.py
@@ -224,7 +224,7 @@ def merge_index(*args, **kwargs):  # pyright: ignore
     raise ValueError(f'Invalid arguments to merge_index: {args}, {kwargs}')
 
 
-def _merge_index_from_list(index_file_urls: List[Any],
+def _merge_index_from_list(index_file_urls: List[Union[str, Tuple[str, str]]],
                            out: Union[str, Tuple[str, str]],
                            keep_local: bool = True,
                            download_timeout: int = 60) -> None:
@@ -281,8 +281,8 @@ def _merge_index_from_list(index_file_urls: List[Any],
             scheme, bucket, path = obj.scheme, obj.netloc, obj.path
             if scheme == '' and bucket == '' and path == '':
                 raise FileNotFoundError(
-                    f'Check data availability! local index {url[0]} is not accessible. remote index {url[1]} does not have a valid url format'
-                )
+                    f'Check data availability! local index {url[0]} is not accessible.' +
+                    'remote index {url[1]} does not have a valid url format')
             dest = os.path.join(temp_root, path.lstrip('/'))
 
             try:
@@ -290,7 +290,7 @@ def _merge_index_from_list(index_file_urls: List[Any],
             except Exception as ex:
                 raise RuntimeError(f'Failed to download index.json: {src} to {dest}') from ex
 
-            if not (os.path.exists(dest)):
+            if not os.path.exists(dest):
                 raise FileNotFoundError(f'Index file {dest} does not exist or not accessible.')
 
             partitions.append(dest)
@@ -334,18 +334,24 @@ def _merge_index_from_root(out: Union[str, Tuple[str, str]], keep_local: bool = 
 
     Args:
         out (Union[str, Tuple[str,str]]): folder that contain MDS partitions.
-            :a local directory, merge index happens locally
-            :a remote directory, download all the sub-directories index.json in a temporary sub-directories, merge locally, and then upload it to out location
-            :a (local_dir, remote_dir), check if sub-directories index.json file present locally
+            :A local directory, merge index happens locally
+            :A remote directory, download all the sub-directories index.json in a temporary sub-directories, merge locally, and then upload it to out location
+            :A (local_dir, remote_dir), check if sub-directories index.json file present locally
                 If yes, then merge locally and upload to remote_dir .
                 If not, download all the sub-directories index.json from remote to local , merge locally, and upload to remote_dir .
         keep_local (bool): Keep local copy of the merged index file. Defaults to ``True``
-        download_timeout (int): The allowed time for downloading each json file. Defaults to 60s.
     """
     from streaming.base.storage.upload import CloudUploader
 
     def not_merged_index(index_file_path: str, out: str):
-        """Check if index_file_path is the merged index at folder out."""
+        """Check if index_file_path is the merged index at folder out.
+
+        Args:
+            index_file_path (str): the path to index.json file
+            out (str): remote or local url of a folder
+        Return:
+            (bool): no if index.json sits in out instead of in the subfolders of out
+        """
         prefix = str(urllib.parse.urlparse(out).path)
         return os.path.dirname(index_file_path).strip('/') != prefix.strip('/')
 
@@ -357,16 +363,16 @@ def _merge_index_from_root(out: Union[str, Tuple[str, str]], keep_local: bool = 
 
     local_index_files = []
     cl = CloudUploader.get(cu.local, exist_ok=True, keep_local=True)
-    for o in cl.list_objects():
-        if o.endswith('.json') and not_merged_index(o, cu.local):
-            local_index_files.append(o)
+    for file in cl.list_objects():
+        if file.endswith('.json') and not_merged_index(file, cu.local):
+            local_index_files.append(file)
 
     if cu.remote:
         obj = urllib.parse.urlparse(cu.remote)
         remote_index_files = []
-        for o in cu.list_objects():
-            if o.endswith(get_index_basename()) and not_merged_index(o, cu.remote):
-                remote_index_files.append(obj.scheme + '://' + os.path.join(obj.netloc, o))
+        for file in cu.list_objects():
+            if file.endswith(get_index_basename()) and not_merged_index(file, cu.remote):
+                remote_index_files.append(obj.scheme + '://' + os.path.join(obj.netloc, file))
         if len(local_index_files) == len(remote_index_files):
             _merge_index_from_list(list(zip(local_index_files, remote_index_files)),
                                    out,
